Moved some files, removed some cruft
