+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "substrate-node-template"
+name = "template-node"
 version = "0.9.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 build = "build.rs"
@@ -10,21 +10,33 @@ path = "src/main.rs"
 
 [dependencies]
 error-chain = "0.12"
-node-cli = { path = "cli" }
 futures = "0.1"
 ctrlc = { version = "3.0", features = ["termination"] }
+log = "0.4"
+tokio = "0.1.7"
+exit-future = "0.1"
+parking_lot = "0.4"
+hex-literal = "0.1"
+slog = "^2"
+parity-codec = { version = "2.1" }
+trie-root = { git = "https://github.com/paritytech/trie" }
+sr-io = { git = "https://github.com/paritytech/substrate" }
+sr-primitives = { git = "https://github.com/paritytech/substrate" }
+substrate-cli = { git = "https://github.com/paritytech/substrate" }
+substrate-primitives = { git = "https://github.com/paritytech/substrate" }
+substrate-executor = { git = "https://github.com/paritytech/substrate" }
+substrate-service = { git = "https://github.com/paritytech/substrate" }
+substrate-transaction-pool = { git = "https://github.com/paritytech/substrate" }
+substrate-network = { git = "https://github.com/paritytech/substrate" }
+substrate-consensus-aura = { git = "https://github.com/paritytech/substrate" }
+template-node-runtime = { path = "runtime" }
 
 [build-dependencies]
 vergen = "2"
 
 [workspace]
-members = [
-	"cli",
-	"runtime"
-]
-exclude = [
-	"runtime/wasm"
-]
+members = [ "runtime" ]
+exclude = [ "runtime/wasm" ]
 
 [profile.release]
 # Substrate runtime requires unwinding.
