+++ README.md
@@ -1,2 +1,3 @@
-# substrate-node-template
-A new SRML-based Substrate node, ready for hacking
+# Template Node
+
+A new SRML-based Substrate node, ready for hacking.
