--- cli/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "node-cli"
-version = "0.9.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-description = "Substrate node implementation in Rust."
-
-[dependencies]
-log = "0.4"
-tokio = "0.1.7"
-exit-future = "0.1"
-parking_lot = "0.4"
-hex-literal = "0.1"
-slog = "^2"
-parity-codec = { version = "2.1" }
-trie-root = { git = "https://github.com/paritytech/trie" }
-sr-io = { git = "https://github.com/paritytech/substrate" }
-sr-primitives = { git = "https://github.com/paritytech/substrate" }
-substrate-cli = { git = "https://github.com/paritytech/substrate" }
-substrate-primitives = { git = "https://github.com/paritytech/substrate" }
-substrate-executor = { git = "https://github.com/paritytech/substrate" }
-substrate-service = { git = "https://github.com/paritytech/substrate" }
-substrate-transaction-pool = { git = "https://github.com/paritytech/substrate" }
-substrate-network = { git = "https://github.com/paritytech/substrate" }
-substrate-consensus-aura = { git = "https://github.com/paritytech/substrate" }
-node-runtime = { path = "../runtime" }
