--- cli/src/lib.rs
@@ -1,112 +1,0 @@
-//! Substrate Node Template CLI library.
-
-#![warn(missing_docs)]
-#![warn(unused_extern_crates)]
-
-extern crate tokio;
-extern crate exit_future;
-#[macro_use]
-extern crate log;
-extern crate substrate_cli as cli;
-extern crate substrate_primitives as primitives;
-extern crate substrate_consensus_aura as consensus;
-#[macro_use]
-extern crate substrate_network as network;
-#[macro_use]
-extern crate substrate_executor;
-extern crate substrate_transaction_pool as transaction_pool;
-#[macro_use]
-extern crate substrate_service;
-extern crate node_runtime;
-
-pub use cli::error;
-mod chain_spec;
-mod service;
-
-use tokio::runtime::Runtime;
-pub use cli::{VersionInfo, IntoExit};
-use substrate_service::{ServiceFactory, Roles as ServiceRoles};
-
-/// The chain specification option.
-#[derive(Clone, Debug)]
-pub enum ChainSpec {
-	/// Whatever the current runtime is, with just Alice as an auth.
-	Development,
-	/// Whatever the current runtime is, with simple Alice/Bob auths.
-	LocalTestnet,
-}
-
-/// Get a chain config from a spec setting.
-impl ChainSpec {
-	pub(crate) fn load(self) -> Result<chain_spec::ChainSpec, String> {
-		Ok(match self {
-			ChainSpec::Development => chain_spec::development_config(),
-			ChainSpec::LocalTestnet => chain_spec::local_testnet_config(),
-		})
-	}
-
-	pub(crate) fn from(s: &str) -> Option<Self> {
-		match s {
-			"dev" => Some(ChainSpec::Development),
-			"local" => Some(ChainSpec::LocalTestnet),
-			_ => None,
-		}
-	}
-}
-
-fn load_spec(id: &str) -> Result<Option<chain_spec::ChainSpec>, String> {
-	Ok(match ChainSpec::from(id) {
-		Some(spec) => Some(spec.load()?),
-		None => None,
-	})
-}
-
-/// Parse command line arguments into service configuration.
-pub fn run<I, T, E>(args: I, exit: E, version: cli::VersionInfo) -> error::Result<()> where
-	I: IntoIterator<Item = T>,
-	T: Into<std::ffi::OsString> + Clone,
-	E: IntoExit,
-{
-	let description = version.description.clone();
-	let executable_name = version.executable_name.clone();
-	let author = version.author.clone();
-
-	match cli::prepare_execution::<service::Factory, _, _, _, _>(args, exit, version, load_spec, &executable_name)? {
-		cli::Action::ExecutedInternally => (),
-		cli::Action::RunService((config, exit)) => {
-			info!("{} ({})", description, executable_name);
-			info!("  version {}", config.full_version());
-			info!("  by {}, 2018", author);
-			info!("  powered by Substrate");
-			info!("Chain specification: {}", config.chain_spec.name());
-			info!("Node name: {}", config.name);
-			info!("Roles: {:?}", config.roles);
-			let mut runtime = Runtime::new()?;
-			let executor = runtime.executor();
-			match config.roles == ServiceRoles::LIGHT {
-				true => run_until_exit(&mut runtime, service::Factory::new_light(config, executor)?, exit)?,
-				false => run_until_exit(&mut runtime, service::Factory::new_full(config, executor)?, exit)?,
-			}
-		}
-	}
-	Ok(())
-}
-
-fn run_until_exit<C, E>(
-	runtime: &mut Runtime,
-	service: service::Service<C>,
-	e: E,
-) -> error::Result<()>
-	where
-		C: substrate_service::Components,
-		E: IntoExit,
-{
-	let (exit_send, exit) = exit_future::signal();
-
-	let executor = runtime.executor();
-	cli::informant::start(&service, exit.clone(), executor.clone());
-
-	let _ = runtime.block_on(e.into_exit());
-	exit_send.fire();
-	Ok(())
-}
