+++ runtime/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "node-runtime"
+name = "template-node-runtime"
 version = "0.9.0"
 authors = ["Parity Technologies <admin@parity.io>"]
