+++ runtime/wasm/build.sh
@@ -7,7 +7,7 @@ else
 	CARGO_CMD="cargo +nightly"
 fi
 $CARGO_CMD build --target=wasm32-unknown-unknown --release
-for i in node_runtime
+for i in template_node_runtime
 do
 	wasm-gc target/wasm32-unknown-unknown/release/$i.wasm target/wasm32-unknown-unknown/release/$i.compact.wasm
 done
