+++ src/main.rs
@@ -1,48 +1,41 @@
-//! Substrate CLI
+//! Substrate Node Template CLI library.
 
 #![warn(missing_docs)]
+#![warn(unused_extern_crates)]
 
-extern crate node_cli as cli;
-extern crate ctrlc;
 extern crate futures;
-
 #[macro_use]
 extern crate error_chain;
+extern crate tokio;
+#[macro_use]
+extern crate log;
+extern crate substrate_cli;
+extern crate substrate_primitives as primitives;
+extern crate substrate_consensus_aura as consensus;
+#[macro_use]
+extern crate substrate_network as network;
+#[macro_use]
+extern crate substrate_executor;
+extern crate substrate_transaction_pool as transaction_pool;
+#[macro_use]
+extern crate substrate_service;
+extern crate template_node_runtime;
 
-use cli::VersionInfo;
-use futures::sync::oneshot;
-use futures::{future, Future};
+mod chain_spec;
+mod service;
+mod cli;
 
-use std::cell::RefCell;
-
-// handles ctrl-c
-struct Exit;
-impl cli::IntoExit for Exit {
-	type Exit = future::MapErr<oneshot::Receiver<()>, fn(oneshot::Canceled) -> ()>;
-	fn into_exit(self) -> Self::Exit {
-		// can't use signal directly here because CtrlC takes only `Fn`.
-		let (exit_send, exit) = oneshot::channel();
-
-		let exit_send_cell = RefCell::new(Some(exit_send));
-		ctrlc::set_handler(move || {
-			if let Some(exit_send) = exit_send_cell.try_borrow_mut().expect("signal handler not reentrant; qed").take() {
-				exit_send.send(()).expect("Error sending exit notification");
-			}
-		}).expect("Error setting Ctrl-C handler");
-
-		exit.map_err(drop)
-	}
-}
-
-quick_main!(run);
+pub use substrate_cli::{VersionInfo, IntoExit, error};
 
 fn run() -> cli::error::Result<()> {
 	let version = VersionInfo {
 		commit: env!("VERGEN_SHA_SHORT"),
 		version: env!("CARGO_PKG_VERSION"),
-		executable_name: "node",
+		executable_name: "template-node",
 		author: "Anonymous <anon@ymous.com>",
-		description: "Substrate Node Template",
+		description: "Template Node",
 	};
-	cli::run(::std::env::args(), Exit, version)
+	cli::run(::std::env::args(), cli::Exit, version)
 }
+
+quick_main!(run);
